LayoutTests/imported/w3c/web-platform-tests/css/cssom-view/scrollingElement-expected.txt


FAIL scrollingElement in quirks mode assert_equals: expected Element node <body style="overflow: clip;"></body> but got null
PASS scrollingElement in no-quirks mode
FAIL scrollingElement in quirks mode (body table) assert_equals: expected null but got Element node <body style="display: table; overflow: auto;"></body>
PASS scrollingElement in no-quirks mode  (body table)
FAIL scrollingElement in quirks mode (root table) assert_equals: expected Element node <body style="overflow: clip;"></body> but got null
PASS scrollingElement in no-quirks mode  (root table)
FAIL scrollingElement in quirks mode (root table, body table) assert_equals: expected null but got Element node <body style="display: table; overflow: auto;"></body>
PASS scrollingElement in no-quirks mode  (root table, body table)
